+++ querylist/__init__.py
@@ -3,7 +3,7 @@
 from querylist import QueryList
 from betterdict import BetterDict
 
-__version__ = "0.0.1"
+__version__ = "0.1.0"
 __author__ = "Thomas Welfley"
 
 __all__ = ['QueryList', 'BetterDict']