+++ docs/source/changelog.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+**0.1.0**
+
+* Renamed QueryList's 'limit' method to 'filter' so that the QueryList API is
+  more consistent with Django's QuerySets.
+
 **0.0.1**
 
 * Initial relase